+++ pyvela/tests/test_data_files.py
@@ -106,6 +106,10 @@ def test_data(model_and_toas):
 
     assert all(np.isfinite(spnta.scaled_toa_unceritainties(spnta.maxlike_params)))
 
+    if spnta.is_wideband():
+        assert all(np.isfinite(spnta.dm_residuals(spnta.maxlike_params)))
+        assert all(np.isfinite(spnta.scaled_dm_unceritainties(spnta.maxlike_params)))
+
     assert all(np.isfinite(spnta.model_dm(spnta.maxlike_params)))
 
     assert (