+++ CHANGELOG.md
@@ -1,5 +1,7 @@
 # Unreleased
 ## Added
+- Use `HDFBackend` in `pyvela` script
+- `scaled_dm_uncertainties()` and `dm_residuals()` methods in `SPNTA` class
 ## Changed
 ## Fixed
 ## Removed